+++ src/task.js
@@ -1,7 +1,13 @@
-export default class Task {
-  constructor(description, completed = false, index) {
+export class Task {
+  constructor(description, index, completed = false) {
     this.description = description;
     this.completed = completed;
     this.index = index;
   }
 }
+
+export const getData = () => JSON.parse(localStorage.getItem('taskslist'));
+
+export const saveData = (tasksList) => {
+  localStorage.setItem('taskslist', JSON.stringify(tasksList));
+};
